lib/crc32:
Hi,

This is an attempt to resolve all the issues that were left in the last review.
There is one proposed change that is still causing a difference of opinion
which has to do with the form of the loops and their performance on x86 and ppc
This version has the change to the form that runs faster on x86 as an ifdef.

This patch series provides improved performance for computing the crc32
polynomial on common hardware by adding the Slicing-by-8 algorithm to the
existing algorithms already included. The new algorithm is very closely
related to the existing algorithm so the extension requires small changes
to implement.  Additionally it cleans up some warnings in the existing
code and adds a kernel mode optional self test to replace the existing
user mode self test.

A description of the existing and new algorithm is included in
Documentation/crc32.txt.
Signed-off-by: Bob Pearson <rpearson@systemfabricworks.com>
Backported to .32 kernel by faux123